Support for floating-point numbers and more
- numbers with decimals
- support for power and square root calculations

+++ src/main/java/com/kevinhe99/Result.java
@@ -1,5 +1,5 @@
 package com.kevinhe99;
 
-public record Result(int value, boolean valid) {
+public record Result(double value, boolean valid) {
 
 }
